LayoutTests/fast/frames/seamless/seamless-designMode-expected.txt

PASS window.iframe.contentDocument.body.textContent is "PASS"
